requirements.txt
feedparser==6.0.11
anthropic>=0.40.0
python-dotenv>=1.0.0
